stop creating resources on map. create on resource init
this is a bit tricky due to dynamic resources (particles). those can have the same resource handle but different vertex/index counts every frame. for now, any resource that has a map_discard, treat as dynamic and build a blas for the entire buffer, instead of building for each hashed draw.

now instead of re-creating resources on map, we mark them as needing a rebuild